SOL-150799: move JWKS-refresh timeout tests to test/integration/
Test_OIDCJWKSRefreshTimeout and Test_OIDCJWKSRefreshTimeout_SSLCertFile
exercise auth.NewAuthMiddleware end-to-end against a fake HTTPS IdP
whose /jwks endpoint hangs. They compose internal/auth's middleware,
the OIDC token verifier, and the go-oidc library — multi-component
behavior that the test/integration/README.md convention says belongs
in this tier rather than next to a single component's unit tests.

Pure relocation. The tests already built their own *http.Client via
idpclient.NewHTTPClient(idpclient.WithTimeout(500ms)) in the previous
commit, so the move is mechanical — they call the same exported
auth.NewAuthMiddleware and idpclient.NewHTTPClient surfaces from
package integration that they used to call from package auth.

Took the relocation as an opportunity to lean on the standard library
where the original code reached past it: server.Certificate() in place
of dereferencing server.TLS.Certificates[0].Certificate[0] and calling
x509.ParseCertificate, and map[string]string in place of
map[string]interface{} for the OIDC discovery JSON.
